+++ src/main/java/com/duriyou/lean/web/UserApiController.java
@@ -1,9 +1,9 @@
 package com.duriyou.lean.web;
 
 import com.duriyou.lean.service.users.UsersService;
-import com.duriyou.lean.web.dto.UsersResponseDto;
-import com.duriyou.lean.web.dto.UsersSaveRequestDto;
-import com.duriyou.lean.web.dto.UsersUpdateRequestDto;
+import com.duriyou.lean.web.dto.Users.UsersResponseDto;
+import com.duriyou.lean.web.dto.Users.UsersSaveRequestDto;
+import com.duriyou.lean.web.dto.Users.UsersUpdateRequestDto;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.*;
