+++ src/main/java/com/duriyou/lean/web/dto/Users/UsersSaveRequestDto.java
@@ -1,4 +1,4 @@
-package com.duriyou.lean.web.dto;
+package com.duriyou.lean.web.dto.Users;
 
 
 import com.duriyou.lean.domain.colleges.Colleges;
