+++ src/main/java/com/duriyou/lean/web/dto/Users/UsersUpdateRequestDto.java
@@ -1,4 +1,4 @@
-package com.duriyou.lean.web.dto;
+package com.duriyou.lean.web.dto.Users;
 
 import com.duriyou.lean.domain.colleges.Colleges;
 import lombok.Builder;
